requirements.txt
a2wsgi==1.10.10
adapters==1.2.0
aiosmtplib==4.0.2
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
apache-airflow==3.0.6
apache-airflow-core==3.0.6
apache-airflow-providers-common-compat==1.7.4
apache-airflow-providers-common-io==1.6.3
apache-airflow-providers-common-sql==1.28.1
apache-airflow-providers-smtp==2.2.1
apache-airflow-providers-standard==1.8.0
apache-airflow-task-sdk==1.0.6
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.9.2
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beartype==0.22.2
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cached-property==2.0.1
cachetools==5.5.2
cadwyn==5.4.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chevron==0.14.0
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.3
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==42.0.8
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.1.1
distro==1.9.0
dnspython==2.8.0
docopt==0.6.2
email-validator==2.3.0
executing==2.2.0
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
fastjsonschema==2.21.2
filelock==3.19.1
Flask==3.1.2
fqdn==1.5.1
frozendict==2.4.6
fsspec==2025.9.0
genanki==0.13.1
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.31.0
googleapis-common-protos==1.70.0
greenback==1.2.1
greenlet==3.2.4
grpcio==1.65.5
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.4.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jaxtyping==0.3.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
lark==1.2.2
lazy-object-proxy==1.12.0
libcst==1.8.4
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
methodtools==0.4.7
mistune==3.1.3
more-itertools==10.8.0
mosestokenizer==1.2.1
mpmath==1.3.0
msgspec==0.19.0
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.2
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.3.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
openai==1.101.0
openfile==0.0.7
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pendulum==3.1.0
pexpect==4.9.0
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
pluggy==1.6.0
prettytable==3.16.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.10.1
PyMuPDF==1.26.4
pyparsing==3.2.3
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
retryhttp==1.3.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rich-argparse==1.7.1
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
Send2Trash==1.8.3
sentry-sdk==2.39.0
setproctitle==1.3.7
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
skops==0.13.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.0
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.48.0
structlog==25.4.0
svcs==25.1.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.21.4
toolwrapper==2.1.0
torch==2.8.0+cu126
torchvision==0.23.0+cu126
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.4.0
typeguard==4.4.4
typer==0.19.2
types-python-dateutil==2.9.0.20250822
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
uctools==1.3.0
universal_pathlib==0.2.6
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.37.0
uvloop==0.21.0
wadler_lindig==0.1.7
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wirerope==1.0.0
wrapt==1.17.3
wtpsplit==2.1.6
youtube-transcript-api==1.2.2
yt-dlp==2025.9.5
zipp==3.23.0
